Add variations to known tokens
I've found that matching the tokens literally did not always yield the
expected results; from time to time, I still write a `TODO:` or even
`:TODO:` which Fixit did not pick up.

In this change we introduce a function for generating some basic
variations of a list of token to include those.

Closes #2.