Using literal tokens
Simplest thing that could possibly work, but not happy with the escaping
of special chars and the escaping of the escape chars, it's just _ugly_,
but it does the trick for now.

Maybe some sort of mapping might be useful, where we have different base
fixit types, for instance 'fixme', 'note' and 'todo' and then each of
them points to a list of the different possible tokens that denote that
type.  This might re-introduce the `build_token_variations` :-D

Closes #6