Introducing token types
Instead of taking the token itself as module in the qf window, we
introduce the notion of token types and show that instead.  This makes
the qf window output more uniform instead of having a mix of slightly
different qf 'modules', so instead of having for instance  'TODO',
'TODO:' and '@todo', there'll be just 3 times 'TODO'.  Which looks a lot
less messy.